renamed ButtonComponents and importet children
