+++ pages/TripDetailPage/index.js
@@ -2,7 +2,7 @@ import { useState } from "react";
 import DestinationImage from "@/components/DestinationImage";
 import TravelOverview from "@/components/TravelDetails";
 import { styled } from "styled-components";
-import Button from "@/components/Button";
+import CircularButton from "@/components/BackButton";
 
 export default function HomePage() {
   return (
@@ -10,7 +10,7 @@ export default function HomePage() {
       <h2>Travel details</h2>
       <DestinationImage />
       <TravelOverview />
-      <Button />
+      <CircularButton>Back to overview</CircularButton>
     </FlexContainer>
   );
 }
